Fix conversion warning

+++ include/boost/decimal/charconv.hpp
@@ -677,7 +677,7 @@ BOOST_DECIMAL_CONSTEXPR auto to_chars_hex_impl(char* first, char* last, const Ta
     // Strip zeros of the significand since frexp10 normalizes it
     const auto zero_removal {detail::remove_trailing_zeros(significand)};
     significand = zero_removal.trimmed_number;
-    exp += zero_removal.number_of_removed_zeros;
+    exp += static_cast<int>(zero_removal.number_of_removed_zeros);
 
     // Calculate the number of bytes
     constexpr auto significand_bits = std::is_same<Unsigned_Integer, std::uint64_t>::value ? 64 : 128;
